==> 07_linear_models/04_statistical_inference_of_stock_returns_with_statsmodels.py <==
#!/usr/bin/env python 
# coding: utf-8

# # Statistical inference of stock returns with linear regression

# ## Imports & Settings

# In[1]:


import warnings
warnings.filterwarnings('ignore')


# In[2]:


get_ipython().run_line_magic('matplotlib', 'inline')

import pandas as pd

from statsmodels.api import OLS, add_constant, graphics
from statsmodels.graphics.tsaplots import plot_acf
from scipy.stats import norm

import seaborn as sns
import matplotlib.pyplot as plt


# In[3]:


sns.set_style('whitegrid')
idx = pd.IndexSlice


# ## Load Data

# In[4]:


with pd.HDFStore('data.h5') as store:
    data = (store['model_data']
            .dropna()
            .drop(['open', 'close', 'low', 'high'], axis=1))


# ### Select Investment Universe

# In[5]:


data = data[data.dollar_vol_rank<100]


# In[6]:


data.info(null_counts=True)


# ### Create Model Data

# In[7]:


y = data.filter(like='target')
X = data.drop(y.columns, axis=1)
X = X.drop(['dollar_vol', 'dollar_vol_rank', 'volume', 'consumer_durables'], axis=1)


# ## Explore Data

# In[8]:


sns.clustermap(y.corr(), cmap=sns.diverging_palette(h_neg=20, h_pos=220), center=0, annot=True, fmt='.2%');


# In[9]:


sns.clustermap(X.corr(), cmap=sns.diverging_palette(h_neg=20, h_pos=220), center=0);
plt.gcf().set_size_inches((14, 14))


# In[10]:


corr_mat = X.corr().stack().reset_index()
corr_mat.columns=['var1', 'var2', 'corr']
corr_mat = corr_mat[corr_mat.var1!=corr_mat.var2].sort_values(by='corr', ascending=False)


# In[11]:


corr_mat.head().append(corr_mat.tail())


# In[12]:


y.boxplot();


# ## Linear Regression for Statistical Inference: OLS with statsmodels

# ### Ticker-wise standardization

# `statsmodels` warns of high design matrix condition numbers. This can arise when the variables are not standardized and the Eigenvalues differ due to scaling. The following step avoids this warning.

# In[13]:


sectors = X.iloc[:, -10:]
X = (X.drop(sectors.columns, axis=1)
     .groupby(level='ticker')
     .transform(lambda x: (x - x.mean()) / x.std())
    .join(sectors)
    .fillna(0))


# ### 1-Day Returns

# In[14]:


target = 'target_1d'
model = OLS(endog=y[target], exog=add_constant(X))
trained_model = model.fit()
print(trained_model.summary())


# ### 5-Day Returns

# In[21]:


target = 'target_5d'
model = OLS(endog=y[target], exog=add_constant(X))
trained_model = model.fit()
print(trained_model.summary())


# #### Obtain the residuals

# In[22]:


preds = trained_model.predict(add_constant(X))
residuals = y[target] - preds


# In[23]:


fig, axes = plt.subplots(ncols=2, figsize=(14,4))
sns.distplot(residuals, fit=norm, ax=axes[0], axlabel='Residuals', label='Residuals')
axes[0].set_title('Residual Distribution')
axes[0].legend()
plot_acf(residuals, lags=10, zero=False, ax=axes[1], title='Residual Autocorrelation')
axes[1].set_xlabel('Lags')
sns.despine()
fig.tight_layout();


# ### 10-Day Returns

# In[19]:


target = 'target_10d'
model = OLS(endog=y[target], exog=add_constant(X))
trained_model = model.fit()
print(trained_model.summary())


# ### Monthly Returns

# In[20]:


target = 'target_21d'
model = OLS(endog=y[target], exog=add_constant(X))
trained_model = model.fit()
print(trained_model.summary())
